Update finance_class.py
Upgraded as far as my knowlege will take me. Maybe want to look into cashe because my goal is for my code to remeber the prior values each time I run the program again. Ex: If i get a dividend, then my cost and number of shares will increase and I want the program to remember that change. I also hae an issue with the find_worth function becuase it's not calculating the new worth when I update variables.